+++ ranked.h
@@ -4,11 +4,11 @@
 #include "votes.h"
 
 typedef struct winner {
-    unsigned int index;
+    uint32_t index;
     mpq_t votes;
 } winner_t;
 
-unsigned int* count_stv(electoral_system_t vote_sys, int num_cands, full_vote_t votes[], int total_votes);
-unsigned int* count_irv(int num_cands, full_vote_t votes[], int num_votes);
+uint32_t* count_stv(electoral_system_t vote_sys, uint32_t num_cands, full_vote_t votes[], uint64_t total_votes);
+uint32_t* count_irv(uint32_t num_cands, full_vote_t votes[], uint64_t num_votes);
 
 #endif /* end of include guard */
