+++ votes.h
@@ -22,39 +22,39 @@ typedef enum voting_method {
 
 // a full preference vote, with all candidates
 typedef struct full_vote {
-    unsigned int* cands;
-    unsigned int cur;
-    unsigned int num_cands;
+    uint32_t* cands;
+    uint32_t cur;
+    uint32_t num_cands;
     mpq_t value;
 } full_vote_t;
 
 // a partial vote for counting, with the current candidate and value
 typedef struct counting_vote {
-    unsigned int cand;
+    uint32_t cand;
     mpq_t value;
 } counting_vote_t;
 
 typedef struct candidate {
     cand_status_t status;
-    int location;
+    uint32_t location;
 } cand_t;
 
 typedef struct list_candidate {
-    int votes;
-    int seats;
+    uint64_t votes;
+    uint32_t seats;
 } list_cand_t;
 
 typedef struct electoral_system {
     voting_method_t method;
-    int winners;
-    int threshold;
+    uint32_t winners;
+    uint32_t threshold;
 } electoral_system_t;
 
-unsigned int find_max_dbl(double count[], int num_cands);
+uint32_t find_max_dbl(double count[], uint32_t num_cands);
 
-unsigned int find_max_int(unsigned int count[], int num_cands, unsigned int threshold);
+uint32_t find_max_int(uint64_t count[], uint32_t num_cands, uint32_t threshold);
 
-unsigned int* count_votes(electoral_system_t vote_sys, cand_t cands[], int num_cands, full_vote_t votes[], uint64_t num_votes);
+uint32_t* count_votes(electoral_system_t vote_sys, cand_t cands[], uint32_t num_cands, full_vote_t votes[], uint64_t num_votes);
 
 counting_vote_t vote_create(full_vote_t full);
